added JWT extended

+++ resources/item.py
@@ -1,5 +1,12 @@
+from flask_jwt_extended import (  # from flask_jwt import jwt_required
+    fresh_jwt_required,
+    get_jwt_claims,
+    get_jwt_identity,
+    jwt_optional,
+    jwt_required,
+)
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+
 # import sqlite3
 from models.item import ItemModel
 
@@ -19,35 +26,37 @@ class Item(Resource):
         help="Every item needs a store id",
     )
 
-
     # @app.route("/student/<string:name>")
-    @jwt_required()  # need access token
+    @jwt_required  # need access token
     def get(self, name):
-#         item = next(filter(lambda x: x["name"] == name, items), None)
-#         if item is not None:
-#             return {"item": item}, 200 if item else 404
-# #         for item in items:
-# #             if item["name"] == name:
-# #                 return item
-#         return {"item": None}, 404  # don't need to raise an Exception
-# #        return {"student": name}  # don't need to jsonify? Yes, due to flask_restful
+        # item = next(filter(lambda x: x["name"] == name, items), None)
+        # if item is not None:
+        #     return {"item": item}, 200 if item else 404
+        # # for item in items:
+        # #     if item["name"] == name:
+        # #         return item
+        # return {"item": None}, 404  # don't need to raise an Exception
+        # # return {"student": name}  # don't need to jsonify? Yes, due to flask_restful
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()  # return item
         return {"message": "Item not found"}, 404
-        
-    def post(self, name):  # has to be same signature as get()
-#         # request_data = request.get_json(force=True)  # do not need content type header
-#         # request_data = request.get_json(silent=True)  # doesn't give error, returns None
-#         if next(filter(lambda x: x["name"] == name, items), None):
-#             return {"message": f"An item with name '{name}' already exists"}, 400  # bad request
 
-# #         request_data = request.get_json()  # separate the name from the price
-#         # price = request_data["price"]
+    @fresh_jwt_required
+    def post(self, name):  # has to be same signature as get()
+        # # request_data = request.get_json(force=True)  # do not need content type header
+        # # request_data = request.get_json(silent=True)  # doesn't give error, returns None
+        # if next(filter(lambda x: x["name"] == name, items), None):
+        #     return {"message": f"An item with name '{name}' already exists"}, 400  # bad request
+
+        # # request_data = request.get_json()  # separate the name from the price
+        # # price = request_data["price"]
         if ItemModel.find_by_name(name):
             return {"message": f"An item with name {name} already exists"}, 400
 
-        data = Item.parser.parse_args()  # how do these functions get the data without arguments?
+        data = (
+            Item.parser.parse_args()
+        )  # how do these functions get the data without arguments?
         # item = {"name": name, "price": data["price"]}
         item = ItemModel(name, **data)
         try:
@@ -55,28 +64,32 @@ class Item(Resource):
             # item.insert()
             item.save_to_db()
         except Exception as e:
-            return {"message": "An error occurred inserting the item"}, 500  # internal server error
+            return {
+                "message": "An error occurred inserting the item"
+            }, 500  # internal server error
 
         return item.json(), 201  # it appears that 200 is default status code
 
     def put(self, name):  # idempotent action
         # data = request.get_json()  # how do these functions get the data without arguments?
-        data = Item.parser.parse_args()  # how do these functions get the data without arguments?
-#         item = next(filter(lambda x: x["name"] == name, items), None)
+        data = (
+            Item.parser.parse_args()
+        )  # how do these functions get the data without arguments?
+        #         item = next(filter(lambda x: x["name"] == name, items), None)
         item = ItemModel.find_by_name(name)
         # updated_item = {"name": name, "price": data["price"]}
         # updated_item = ItemModel(name, data["price"])
         if item is None:
-#             item = {"name": name, "price": data["price"]}
-#             items.append(item)
+            # item = {"name": name, "price": data["price"]}
+            # items.append(item)
             # try:
             #     # ItemModel.insert(updated_item)
             #     updated_item.insert()
             # except:
             #     return {"message": "An error occurred inserting the item"}, 500
-            item = ItemModel(name, **data) 
+            item = ItemModel(name, **data)
         else:
-#             item.update(data)
+            # item.update(data)
             # try:
             #     # ItemModel.update(updated_item)
             #     updated_item.update()
@@ -88,9 +101,10 @@ class Item(Resource):
         # return updated_item.json()
         return item.json()
 
+    @jwt_required
     def delete(self, name):  # should the delete body be moved to ItemModel? no
-#         global items  # allows "mutating" list
-#         items = list(filter(lambda x: x["name"] != name, items))
+        # global items  # allows "mutating" list
+        # items = list(filter(lambda x: x["name"] != name, items))
 
         # connection = sqlite3.connect("data.db")
         # cursor = connection.cursor()
@@ -99,13 +113,18 @@ class Item(Resource):
         # connection.commit()
         # connection.close()
         # return {"message": "Item deleted"}
+        claims = get_jwt_claims()
+        if not claims["is_admin"]:
+            return {"message": "Admin privilege required."}, 401
 
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
         return {"message": "Item deleted"}
 
+
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
         # connection = sqlite3.connect("data.db")
         # cursor = connection.cursor()
@@ -115,8 +134,15 @@ class ItemList(Resource):
         # for row in result:
         #     items.append({"name": row[0], "price": row[1]})
         # connection.close()
-        
-        # return {"items": items}  # must return a dictionary    
 
-#         return {"items": [item.json() for item in ItemModel.query.all()]}
-        return {"items": [item.json() for item in ItemModel.find_all()]}
+        # return {"items": items}  # must return a dictionary
+
+        # return {"items": [item.json() for item in ItemModel.query.all()]}
+        items = [item.json() for item in ItemModel.find_all()]
+        user_id = get_jwt_identity()  # if user_id is None, then not logged in
+        if user_id:
+            return {"items": items}, 200
+        return {
+            "items": [item["name"] for item in items],
+            "message": "More data available if you log in",
+        }, 200
